+++ src/Service/Sqs/src/SqsClient.php
@@ -382,6 +382,13 @@ class SqsClient extends AbstractApi
                     'signService' => 'sqs',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-east-1':
+                return [
+                    'endpoint' => 'https://sqs.us-east-1.amazonaws.com',
+                    'signRegion' => 'us-east-1',
+                    'signService' => 'sqs',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-gov-east-1':
                 return [
                     'endpoint' => 'https://sqs.us-gov-east-1.amazonaws.com',